requirements.txt
alembic==1.4.3
altgraph==0.17
Babel==2.8.0
baidu-aip==2.2.18.0
blinker==1.4
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
dnspython==2.0.0
dominate==2.6.0
easyhid==0.0.10
email-validator==1.1.2
Flask==1.1.2
Flask-Babel==2.0.0
Flask-Bootstrap==3.3.7.1
Flask-JWT==0.3.2
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==2.5.3
Flask-Moment==0.10.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
future==0.18.2
guess-language-spirit==0.5.3
idna==2.10
importlib-metadata==2.0.0
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
numpy==1.19.3
pefile==2019.4.18
Pillow==8.0.1
py2exe==0.10.0.2
PyAudio==0.2.11
pycparser==2.20
pyinstaller==4.0
pyinstaller-hooks-contrib==2020.10
PyJWT==1.4.2
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
python-vlc==3.0.11115
pytz==2020.4
pywin32==228
pywin32-ctypes==0.2.0
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.20
UNKNOWN==0.0.0
urllib3==1.25.11
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.3
wxPython==4.1.0
zipp==3.4.0
